Move Cache back to common and setup Native impl
