+++ zipline-loader/src/jvmTest/kotlin/app/cash/zipline/loader/ZiplineLoaderTest.kt
@@ -23,8 +23,8 @@ import kotlin.test.BeforeTest
 import kotlin.test.Test
 import kotlin.test.assertEquals
 import kotlinx.coroutines.ExperimentalCoroutinesApi
-import kotlinx.coroutines.test.TestCoroutineDispatcher
-import kotlinx.coroutines.test.runBlockingTest
+import kotlinx.coroutines.test.UnconfinedTestDispatcher
+import kotlinx.coroutines.test.runTest
 import kotlinx.serialization.encodeToString
 import kotlinx.serialization.json.Json
 import okio.Buffer
@@ -39,7 +39,7 @@ import okio.fakefilesystem.FakeFileSystem
 @ExperimentalCoroutinesApi
 class ZiplineLoaderTest {
   private val httpClient = FakeZiplineHttpClient()
-  private val dispatcher = TestCoroutineDispatcher()
+  private val dispatcher = UnconfinedTestDispatcher()
   private val driver = JdbcSqliteDriver(JdbcSqliteDriver.IN_MEMORY)
   private val cacheSize = 64
   private var nowMillis = 1_000L
@@ -71,15 +71,13 @@ class ZiplineLoaderTest {
   }
 
   @Test
-  fun happyPath() {
+  fun happyPath() = runTest {
     httpClient.filePathToByteString = mapOf(
       alphaFilePath to alphaBytecode(quickJs),
       bravoFilePath to bravoBytecode(quickJs)
     )
     val zipline = Zipline.create(dispatcher)
-    dispatcher.runBlockingTest {
-      loader.load(zipline, manifest(quickJs))
-    }
+    loader.load(zipline, manifest(quickJs))
     assertEquals(
       zipline.quickJs.evaluate("globalThis.log", "assert.js"),
       """
@@ -90,16 +88,14 @@ class ZiplineLoaderTest {
   }
 
   @Test
-  fun loadManifestFromUrl() {
+  fun loadManifestFromUrl() = runTest {
     httpClient.filePathToByteString = mapOf(
       manifestPath to Json.encodeToString(manifest(quickJs)).encodeUtf8(),
       alphaFilePath to alphaBytecode(quickJs),
       bravoFilePath to bravoBytecode(quickJs)
     )
     val zipline = Zipline.create(dispatcher)
-    dispatcher.runBlockingTest {
-      loader.load(zipline, manifestPath)
-    }
+    loader.load(zipline, manifestPath)
     assertEquals(
       zipline.quickJs.evaluate("globalThis.log", "assert.js"),
       """
